requirements.txt
Flask
python-dotenv
boto3
django
djangorestframework